Hi Thomas,
For the GEOM & SMESH modules it is really important that python lists can be converted to std::list.
This simple example of the sketcher does point out this limitation.

What could be a constructive approach is that to look in how Salome does this swigging.
Also, Salome excepts a standard python string rather than a TCollection_AsciiString.

So, in a way Salome is more pythonic than PythonOCC ;')
Let's see if we can do better!
It would be great if TCol* TCollection* could be automagically be adapted...
I suppose its possible in SWIG?

A+

-jelle