Initial commit of the refactored swig generators, sphinx-documentation generator. should generate decent swig files, but havent tested compilation on the last revisions cd to refactored_wrapper directory and run build.py
documentation generator is not working, as it depends on a django/sqlite database model that isnt working. 

mainly just for reference, lets see how and if we can pick out useful bits and integrate it in the current wrapper

Main differences from the current swig generator is that it tries to use/extend the pygccxml datamodel as much as possible and separate the parsing step from the writing step, so filters out all symbols that should not be wrapped, moves the classes into namespaces and adds flags and other meta data (renames, special directives etc). 
in the 2nd step it uses a simple template system to generate the swig files.