Improved SimpleGui, added Tkinter support
loading the backends was handled in a really silly way. for instance, in the
previous version _all_ these backend would get loaded. serious overhead... so,
pythonocc is now more selective about loading the backend and does so in a
logical order.

fixed the menu bar issue, and added a method that centers the window and sizes the
window, depending on the resolution of the screen.

Also adds an optional argument to init_display to overload the default backend.